Cleanup after CR
