+++ .github/CODEOWNERS
@@ -174,6 +174,7 @@ mkdocs.yaml @hmellor
 /vllm/v1/engine/utils.py @skhorasganiTT @viktorpusTT
 /vllm/v1/engine/tt_core_launcher.py @skhorasganiTT @viktorpusTT
 /vllm/v1/engine/tt_dp_gather.py @viktorpusTT @tchedaTT
+/vllm/v1/engine/tt_engine_step.py @viktorpusTT @tchedaTT
 
 # Tests
 /tests/tt @tchedaTT @skhorasganiTT @viktorpusTT
